Initial WebSocket Support (#16)
* Add the .envrc file to the .gitignore

* WIP progress on WebSocket implementation

* WIP progress on WebSocket implementation

* Update integration tests

* Dependency upgrades

* Still working on the intermittent failure

* Try updating dependencies

* Add the associated Show to prevent the inexplicably intermittent test failure

* Revamp the Github actions

* Move to dtolnay

* Iterating on the build

* More build changes

* Move coverage back into ci

* Try to combine unit testing and coverage steps

+++ libs/domains/src/episodes/queries.rs
@@ -18,16 +18,16 @@ pub struct EpisodesPage {
     data: Vec<Episode>,
 
     /// The number of `Episodes` returned for the current page
-    count: usize,
+    count: u64,
 
     /// Tne total number of `Episodes` available
-    total: usize,
+    total: u64,
 
     /// The current page
-    page: usize,
+    page: u64,
 
     /// The number of pages available
-    page_count: usize,
+    page_count: u64,
 }
 
 impl From<ManyResponse<Episode>> for EpisodesPage {
